+++ WattApp/server/Server/Services/IUserService.cs
@@ -5,7 +5,7 @@ namespace Server.Services
 {
     public interface IUserService
     {
-        Task<UserModel?>? GetUserById(int id);
+        Task<UserModel?>? GetUserById(long id);
         Task<UserModel?>? GetUserByEmail(string email);
         Task<UserModel?> GetUserByUsername(string username);
         Task<DataPage<object>> GetPageOfUsers(int pageNumber, int itemsPerPage, Func<UserModel, bool> filter);
